no se renderizan los cambios a la segunda forma de store
