+++ src/components/UsersComponents/BusquedaUser.js
@@ -1,30 +1,31 @@
 import React, { Component } from 'react';
 import  { Link }  from 'react-router-dom';
-import  { Card, AppBar, Toolbar, InputBase, Typography, IconButton, Icon, Button } from '@material-ui/core';
+import  { Table, Card, AppBar, Toolbar, InputBase, Typography, IconButton, Icon, Button } from '@material-ui/core';
 import {  connect } from 'react-redux';
-import { loadingUser,addOnlyUser, findOnlyUser, editOnlyUser, deleteOnlyUser } from './../../../../state/actions/UsersActions/UserOnlyActions';
-import { fetchDependants } from './../../../../state/actions/DependantsActions/DependantsActions';
-import NavBar from '../../NavBar';
-import './BusquedaDependant.scss';
+import { loadingUser,addOnlyUser, findOnlyUser, editOnlyUser, deleteOnlyUser } from './../../../../../state/actions/UsersActions/UserOnlyActions';
+import { fetchUsers } from './../../../../../state/actions/UsersActions/UsersActions';
+import NavBar from '../NavBar';
 
 
 
-class BusquedaDependant extends Component {
+
+
+class BusquedaUser extends Component {
 
     render() {
         return (
-            <div className='busqueda_dependant'>
+            <div className='busqueda_user'>
                 <AppBar  position="static" color="success">
                     <NavBar />
                     <Link to="/Home" className="link"><Button>Home</Button></Link>    
    
                     <Toolbar>
                         <Typography className='title' variant="h6" color="inherit" noWrap>
-                            RED - Busqueda de Dependientes
+                            RED - Busqueda de Usuarios
                         </Typography>
                         <div className=''>
                             <div className=''>
-                                <IconButton>
+                                <IconButton onClick={this.handleAddUser}>
                                     <Icon>
                                         searchicon
                                     </Icon>
@@ -35,9 +36,9 @@ class BusquedaDependant extends Component {
                     </Toolbar>    
                 </AppBar>
                 <Card>
-
-                    Resultado de la Busqueda
-
+                    <Table>
+                        Resultado de la Busqueda
+                    </Table>
                     <IconButton>
                             <Icon>
                                 edit
@@ -56,7 +57,7 @@ class BusquedaDependant extends Component {
 
 const mapDispatchToProps = (dispatch) => {
     return {
-        fetchDependants,
+        fetchUsers,
         loadingUser,
         addOnlyUser,
         findOnlyUser,
@@ -73,4 +74,4 @@ const mapStateToProps = (state) => {
     }
 }
 
-export default connect(mapDispatchToProps,mapStateToProps) (BusquedaDependant);
+export default connect(mapDispatchToProps,mapStateToProps) (BusquedaUser);
