--- src/components/UsersComponents/UsersMenu/UsersSubMenus/EstilosUsersSubMenus/BusquedaUser.scss
@@ -1,20 +1,0 @@
-.busqueda_user {
-
-    searchIcon {
-        width: 9;
-        height: 100%;
-        position: absolute;
-        pointer-events: none;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-
-
-   InputBase {
-        color: darkgray ;
-        width: 100%;
-        border-color: black;
-        
-   }    
-}
